Move old commands to different files
