+++ .editorconfig
@@ -11,7 +11,7 @@ charset = utf-8
 end_of_line = lf
 insert_final_newline = true
 trim_trailing_whitespace = true
-indent_style = tab
+indent_style = space
 
 [{.jshintrc,*.json,*.yml,*.feature}]
 indent_style = space
